chore(analytics-core): move browser client under types
